Create IdrinthFollowerDialogueQuestNamira07.md
First line if the player says: "No. I just want to lure the entire coven in one place, so they will meet justice." 
Second line if the player says :"I'm not feeling well at the thought... but if we don't get the priest, Eola will. He's safer with us."
Third line if the player says: "Where's the harm in a feast? Besides, priests of Arkay hate you, so no loss."
When choosing the third option, Idrinth will leave the player.